feat: inspect html file preview
