+++ apps/cloud/src/app/@shared/files/preview/file-html-preview.component.html
@@ -1,0 +1,10 @@
+@if (htmlPreviewUrl(); as previewUrl) {
+  <iframe
+    #htmlPreviewFrame
+    [src]="previewUrl | safe: 'resourceUrl'"
+    sandbox="allow-downloads allow-forms allow-modals allow-popups allow-scripts"
+    referrerpolicy="no-referrer"
+    class="h-full min-h-0 w-full flex-1 border-none bg-components-panel-bg"
+    (load)="handleHtmlPreviewLoad()"
+  ></iframe>
+}
