+++ apps/cloud/src/app/@shared/files/preview/file-preview-content.component.html
@@ -47,32 +47,18 @@
         <pre class="flex-1 overflow-auto p-4 text-sm whitespace-pre-wrap"><code>{{ content() ?? '' }}</code></pre>
       }
       @case ('html') {
-        @if (htmlPreviewUrl(); as previewUrl) {
-          <div class="flex h-full min-h-0 flex-col">
-            @if (htmlInspectAvailable()) {
-              <div class="flex items-center justify-end border-b border-divider-regular px-3 py-2">
-                <button
-                  type="button"
-                  data-html-inspect-button="true"
-                  [class]="htmlInspectMode() ? 'btn btn-primary btn-small' : 'btn btn-secondary btn-small'"
-                  (click)="toggleHtmlInspectMode()"
-                >
-                  <i class="ri-focus-3-line mr-1"></i>
-                  {{ 'PAC.Chat.ClawXpert.Inspect' | translate: { Default: 'Inspect' } }}
-                </button>
-              </div>
-            }
-
-            <iframe
-              #htmlPreviewFrame
-              [src]="previewUrl | safe: 'resourceUrl'"
-              sandbox="allow-downloads allow-forms allow-modals allow-popups allow-scripts"
-              referrerpolicy="no-referrer"
-              class="min-h-0 flex-1 border-none bg-components-panel-bg"
-              (load)="handleHtmlPreviewLoad()"
-            ></iframe>
-          </div>
-        }
+        <pac-file-html-preview
+          class="h-full"
+          [content]="content()"
+          [filePath]="filePath()"
+          [fileName]="fileName()"
+          [inspectMode]="htmlInspectMode()"
+          [loading]="loading()"
+          [referenceable]="referenceable()"
+          [url]="url()"
+          (fileElementReference)="fileElementReference.emit($event)"
+          (inspectModeChange)="htmlInspectModeChange.emit($event)"
+        />
       }
       @case ('document') {
         <div #documentPreviewHost class="relative h-full overflow-auto" data-file-preview-document-host="true">
