+++ apps/cloud/src/app/@shared/files/viewer/viewer.component.spec.ts
@@ -1,4 +1,6 @@
+import { provideHttpClient } from '@angular/common/http'
 import { TestBed } from '@angular/core/testing'
+import { By } from '@angular/platform-browser'
 import { TranslateModule } from '@ngx-translate/core'
 import { FileViewerComponent, inferMarkdownPreviewSelection } from './viewer.component'
 
@@ -18,15 +20,38 @@ jest.mock('@xpert-ai/ocap-angular/common', () => {
 })
 
 jest.mock('@xpert-ai/headless-ui', () => {
-  const { Component, Input } = jest.requireActual('@angular/core')
+  const { Component, EventEmitter, forwardRef, Input, Output } = jest.requireActual('@angular/core')
+  const { NG_VALUE_ACCESSOR } = jest.requireActual('@angular/forms')
 
   @Component({
     standalone: true,
     selector: 'z-segmented',
-    template: '<ng-content />'
+    template: '<ng-content />',
+    providers: [
+      {
+        provide: NG_VALUE_ACCESSOR,
+        useExisting: forwardRef(() => ZardSegmentedComponent),
+        multi: true
+      }
+    ]
   })
   class ZardSegmentedComponent {
     @Input() ngModel?: unknown
+    @Output() ngModelChange = new EventEmitter()
+    private onChange: (value: unknown) => void = () => undefined
+    private onTouched: () => void = () => undefined
+
+    writeValue(value: unknown) {
+      this.ngModel = value
+    }
+
+    registerOnChange(onChange: (value: unknown) => void) {
+      this.onChange = onChange
+    }
+
+    registerOnTouched(onTouched: () => void) {
+      this.onTouched = onTouched
+    }
   }
 
   @Component({
@@ -47,7 +72,7 @@ jest.mock('@xpert-ai/headless-ui', () => {
 })
 
 jest.mock('ngx-markdown', () => {
-  const { Component, Input } = jest.requireActual('@angular/core')
+  const { Component, Input, NgModule } = jest.requireActual('@angular/core')
 
   @Component({
     standalone: true,
@@ -58,8 +83,14 @@ jest.mock('ngx-markdown', () => {
     @Input() data?: string
   }
 
+  @NgModule({
+    imports: [MarkdownComponent],
+    exports: [MarkdownComponent]
+  })
+  class MarkdownModule {}
+
   return {
-    MarkdownModule: class MarkdownModule {},
+    MarkdownModule,
     MarkdownComponent
   }
 })
@@ -78,7 +109,9 @@ jest.mock('../editor/editor.component', () => {
     @Input() wordWrap?: boolean
     @Input() fileName?: string
     @Input() content?: string
+    @Input() referenceable?: boolean
     @Output() contentChange = new EventEmitter()
+    @Output() referenceSelection = new EventEmitter()
     @Output() selectionChange = new EventEmitter()
   }
 
@@ -107,8 +140,10 @@ jest.mock('../preview/file-preview-content.component', () => {
     @Input() referenceable?: boolean
     @Input() spreadsheet?: unknown
     @Input() url?: string | null
+    @Input() htmlInspectMode?: boolean
     @Output() download = new EventEmitter<void>()
     @Output() fileElementReference = new EventEmitter()
+    @Output() htmlInspectModeChange = new EventEmitter<boolean>()
     @Output() referenceSelection = new EventEmitter()
   }
 
@@ -120,14 +155,9 @@ jest.mock('../preview/file-preview-content.component', () => {
 describe('FileViewerComponent', () => {
   beforeEach(async () => {
     TestBed.resetTestingModule()
-    TestBed.overrideComponent(FileViewerComponent, {
-      set: {
-        template: '',
-        imports: []
-      }
-    })
     await TestBed.configureTestingModule({
-      imports: [TranslateModule.forRoot(), FileViewerComponent]
+      imports: [TranslateModule.forRoot(), FileViewerComponent],
+      providers: [provideHttpClient()]
     }).compileComponents()
   })
 
@@ -206,6 +236,51 @@ describe('FileViewerComponent', () => {
     component.updatePreviewMode('code')
 
     expect(component.htmlPreviewReferenceable()).toBe(false)
+  })
+
+  it('shows the html inspect button in preview mode and passes inspect mode to preview content', () => {
+    const fixture = TestBed.createComponent(FileViewerComponent)
+    fixture.componentRef.setInput('filePath', 'index.html')
+    fixture.componentRef.setInput('content', '<!doctype html><html><body><button>Preview</button></body></html>')
+    fixture.componentRef.setInput('readable', true)
+    fixture.componentRef.setInput('referenceable', true)
+    fixture.detectChanges()
+
+    const component = fixture.componentInstance
+    const button = fixture.debugElement.query(By.css('[data-html-inspect-button="viewer"]'))
+    const preview = fixture.debugElement.query(By.css('pac-file-preview-content'))
+    expect(component.canInspectHtmlPreview()).toBe(true)
+    expect(button).not.toBeNull()
+    expect(preview.componentInstance.htmlInspectMode).toBe(false)
+
+    ;(button.nativeElement as HTMLButtonElement).click()
+    fixture.detectChanges()
+
+    expect(component.htmlInspectMode()).toBe(true)
+    expect((button.nativeElement as HTMLButtonElement).className).toContain('btn-primary')
+    expect(preview.componentInstance.htmlInspectMode).toBe(true)
+
+    preview.componentInstance.htmlInspectModeChange.emit(false)
+    fixture.detectChanges()
+
+    expect(component.htmlInspectMode()).toBe(false)
+
+    component.toggleHtmlInspectMode()
+    expect(component.htmlInspectMode()).toBe(true)
+
+    fixture.componentRef.setInput('filePath', 'other.html')
+    fixture.detectChanges()
+
+    expect(component.htmlInspectMode()).toBe(false)
+
+    component.toggleHtmlInspectMode()
+    expect(component.htmlInspectMode()).toBe(true)
+
+    component.updatePreviewMode('code')
+    fixture.detectChanges()
+
+    expect(component.htmlInspectMode()).toBe(false)
+    expect(fixture.debugElement.query(By.css('[data-html-inspect-button="viewer"]'))).toBeNull()
   })
 
   it('enables the inline selection action for markdown files after switching to edit mode', () => {
